Upgrade flask-cors to 6.0.1 (#147)
This is the result of running `pipenv upgrade flask-cors`. Our Pipfile
supports the correct version; we just need to update the lock file.

This is to fix an issue in earlier versions of `flask-cors`:
CVE-2024-6866. While it doesn't seem to apply to our use case at first
glance, it's easy enough to fix, so seems worth it to do so.